requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
Pillow
opencv-python

# Flask==2.0.1
# Pillow==8.3.1
# numpy==1.19.5
# opencv-python==4.5.3.56
# pandas==1.3.3
# seaborn==0.11.2
# matplotlib==3.4.3
# scikit-learn==0.24.2
# tensorflow==2.15.1
